requirements.txt
PyPDF2==3.0.1
langchain==0.2.11
langchain-google-genai==1.0.8
google-generativeai==0.7.2
langchain-community==0.2.10
python-dotenv==1.0.1
streamlit==1.32.2
faiss-cpu==1.8.0.post1
torch==2.0.1
torchvision==0.15.2
flask_sqlalchemy==3.1.1

absl-py==2.1.0
altair==5.2.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
executing==2.0.1
filelock==3.13.3
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.50.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-pasta==0.2.0
grpcio==1.62.1
gunicorn==21.2.0
h5py==3.10.0
idna==3.6
ipykernel==6.29.4
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pickle-mixin==1.0.2
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==69.2.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.32.2
streamlit-option-menu==0.3.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel
termcolor==2.4.0
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
waitress==3.0.0
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
xgboost==2.0.3
